Add basic noop block placement implementation
