+++ crates/server/src/packet_handlers.rs
@@ -1,6 +1,6 @@
 use base::{Position, Text};
 use common::{chat::ChatKind, Game};
-use digging::handle_player_digging;
+use interaction::{handle_player_block_placement, handle_player_digging};
 use ecs::{Entity, EntityRef, SysResult};
 use protocol::{
     packets::{
@@ -13,7 +13,7 @@ use quill_common::components::Name;
 
 use crate::{NetworkId, Server};
 
-mod digging;
+mod interaction;
 pub mod inventory;
 mod movement;
 
@@ -52,6 +52,8 @@ pub fn handle_packet(
             inventory::handle_click_window(server, player, packet)
         }
 
+        ClientPlayPacket::PlayerBlockPlacement(packet) => handle_player_block_placement(server, packet, player),
+
         ClientPlayPacket::TeleportConfirm(_)
         | ClientPlayPacket::QueryBlockNbt(_)
         | ClientPlayPacket::SetDifficulty(_)
@@ -89,7 +91,6 @@ pub fn handle_packet(
         | ClientPlayPacket::UpdateStructureBlock(_)
         | ClientPlayPacket::UpdateSign(_)
         | ClientPlayPacket::Spectate(_)
-        | ClientPlayPacket::PlayerBlockPlacement(_)
         | ClientPlayPacket::UseItem(_) => Ok(()),
     }
 }
